+++ Tests/WindowsCloudPC.Tests.ps1
@@ -19,6 +19,7 @@ BeforeDiscovery {
         'Get-CloudPCSupportedRegion',
         'Get-CloudPCUsage',
         'Get-CloudPCUserSetting',
+        'Invoke-CloudPCEndGracePeriod',
         'Invoke-CloudPCPolicyReprovision',
         'Invoke-CloudPCReprovision',
         'New-CloudPCMaintenanceWindow',